+++ stories/productionSystem/index.jsx
@@ -11,9 +11,6 @@
 import { storiesOf } from '@storybook/react';
 import { StoryIntro, StoryBody } from '../wrappers';
 
-import ImageSearcher from '../molecules/imageSearch';
-import AudioSearcher from '../molecules/audioSearch';
-import VideoSearcher from '../molecules/videoSearch';
 import FormExampleAuthors from './FormExampleAuthors';
 import HowtoExamples from './HowtoExamples';
 import MultiSelectDropdownExample from './MultiSelectDropdownExample';
@@ -23,39 +20,6 @@ import NdlaFilmEditorExample from './NdlaFilmEditorExample';
 import CodeBlockEditorExample from './CodeBlockEditorExample';
 
 storiesOf('Other/Production system', module)
-  .add('Image search', () => (
-    <div>
-      <StoryIntro title="Bildesøk">
-        <p>Bildesøk som gjør det mulig å søke mot NDLA sitt bilde api.</p>
-      </StoryIntro>
-      <StoryBody>
-        <ImageSearcher />
-      </StoryBody>
-    </div>
-  ))
-  .add('Audio search', () => (
-    <div>
-      <StoryIntro title="Lydsøk">
-        <p>Lydsøk som gjør det mulig å søke mot NDLA sitt lyd-api..</p>
-      </StoryIntro>
-      <StoryBody>
-        <AudioSearcher />
-      </StoryBody>
-    </div>
-  ))
-  .add('Video search', () => (
-    <div>
-      <StoryIntro title="Videosøk">
-        <p>
-          Videsøk som gjør det mulig å søke mot NDLA sine videoer på brightcove. Denne modulen krever at det både finnes
-          en token og api url (for brightcove).
-        </p>
-      </StoryIntro>
-      <StoryBody>
-        <VideoSearcher />
-      </StoryBody>
-    </div>
-  ))
   .add('Form elements', () => (
     <div>
       <StoryIntro title="Form elementer">
